Rename to slack.py
This aligns our ducks, and makes /python load slack.py and /python
reload slack more similar.
